requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
sqlalchemy==2.0.36
flask==3.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
fake-useragent==1.4.0
schedule==1.2.0
